Refresh targets when adding new target
