+++ web/src/components/Targets/TargetContext.tsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const TargetContext = createContext({ refresh: () => {} });
