+++ web/src/components/Targets/index.tsx
@@ -1,0 +1,3 @@
+export * from './CreateTarget';
+export * from './TargetContext';
+export * from './Targets';
