+++ web/src/components/index.tsx
@@ -1,4 +1,4 @@
 export * from './Form';
 export * from './NavBar';
 export * from './Delegations';
-export * from './CreateTarget';
+export * from './Targets';
